requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
mlflow
jupyter
optuna
hydra-core
omegaconf
dvc
joblib
apache-airflow
psycopg2-binary